+++ content/3.community/1.wallet-setup/2.keplr/1.keplr-setup.md
@@ -59,9 +59,16 @@ Warning
 
 ## Add Archway to Keplr
 
+::addToKeplr{:chain="mainnet" :class="mr-2"}
+Add Archway Mainnet
+::
 
-::addToKeplr
-Add Archway
+::addToKeplr{:chain="testnet" :class="mr-2"}
+Add Archway Testnet (Constantine)
+::
+
+::addToKeplr{:chain="devnet"}
+Add Archway Devnet (Titus)
 ::
 
 ### Keplr add more chains
